Load distance and duration in java batch environment
Require small change in DB content.  Also please think about the API
query limit. It is not pratical to dynamic compute distance in the
search algorithm. Possible solution is to use the average limit at a
specific field. This could be a good estimation for the driving time
when another route joined.